WC2-172: refactor StarsComponent
